requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
flask-cors
asgiref
python-multipart
numpy
pydantic